Created userguide.hs containing updated examples
The original 2001 documentation for the Parsec package contained a User Guide, copy of which can be found at http://research.microsoft.com/en-us/um/people/daan/download/parsec/parsec.pdf. The related examples were formerly contained in a parsec/examples/userguide file that is no longer part of the package. Not surprisingly, due to the 15 years since the original release of the User Guide, some of the examples have become outdated due to changes in the library or GHC or both. userguide.hs has updated necessary pragmas and imports (and in some cases, syntax) to reproduce the results show in the User Guide.